pyproject.toml
[tool.poetry]
name = "bike-fitter"
version = "0.1.0"
description = "CLI tool for at-home bike fitting via video analysis"
authors = ["Manuel Fideles"]
maintainers = ["Manuel Fideles"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "3.12.4"
typer = "^0.16.0"
rich = "^14.0.0"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.10"
alembic = "^1.16.2"
opencv-python = "^4.11.0.86"
mediapipe = "^0.10.21"
numpy = "1.26.0"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
